pyproject.toml
[project]
name = "fraud-det"
version = "0.1.0"
description = "Advanced Fraud Detection System using Multi-Agent Systems & Chain of Thought Reasoning"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Navya", email = "navyag3@illinois.edu" }]

dependencies = [
	"fastapi>=0.104.0",
	"uvicorn>=0.24.0",
	"pydantic>=2.4.2",
	"motor>=3.3.1",
	"redis>=5.0.1",
	"networkx>=3.2.1",
	"numpy>=1.26.0",
	"scipy>=1.11.3",
	"scikit-learn>=1.3.2",
	"openai>=1.3.0",
	"python-jose>=3.3.0",
	"passlib>=1.7.4",
	"pyyaml>=6.0.1",
	"aiosmtplib>=2.0.2",
	"twilio>=8.10.0",
	"jinja2>=3.1.2",
	"prometheus-client>=0.17.1",
	"httpx>=0.25.1",
	"tiktoken>=0.5.1",
	"tenacity>=8.2.3",
	"pandas>=2.1.2",
	"torch>=2.1.0",
	"python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
	"pytest>=7.4.3",
	"pytest-asyncio>=0.21.1",
	"black>=23.10.1",
	"flake8>=6.1.0",
	"mypy>=1.6.1",
	"isort>=5.12.0",
	"pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/wonkyvamp/fraud-det"
Documentation = "https://github.com/wonkyvamp/fraud-det#readme"
Repository = "https://github.com/wonkyvamp/fraud-det.git"
Issues = "https://github.com/wonkyvamp/fraud-det/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=fraud_det"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.coverage.run]
source = ["fraud_det"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
	"pragma: no cover",
	"def __repr__",
	"raise NotImplementedError",
	"if __name__ == .__main__.:",
	"pass",
]
